Uploaded the mirin ""package manager""
I hope this gets a laugh out of someone. This was more of a test to see
how effective or not lua was at making command line tools. This tool is a
little bit convenient to have, but also is completely unnecessary.

Thank you for reading this. I hope you have a great day!